+++ src/main/java/com/samourai/whirlpool/client/wallet/orchestrator/AutoTx0Orchestrator.java
@@ -12,6 +12,7 @@ import com.samourai.whirlpool.client.wallet.WhirlpoolWalletConfig;
 import com.samourai.whirlpool.client.wallet.beans.*;
 import com.samourai.whirlpool.client.whirlpool.beans.Pool;
 import java.util.Collection;
+import java.util.List;
 import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -47,7 +48,7 @@ public class AutoTx0Orchestrator extends AbstractOrchestrator {
     }
   }
 
-  public Tx0 autoTx0() throws Exception { // throws AutoMixInsufficientBalanceException
+  public List<Tx0> autoTx0() throws Exception { // throws AutoMixInsufficientBalanceException
     String poolId = config.getAutoTx0PoolId();
     Pool pool = whirlpoolWallet.getPoolSupplier().findPoolById(poolId);
     if (pool == null) {
@@ -61,7 +62,7 @@ public class AutoTx0Orchestrator extends AbstractOrchestrator {
             pool, tx0FeeTarget, mixFeeTarget); // throws AutoMixInsufficientBalanceException
 
     Tx0Config tx0Config = whirlpoolWallet.getTx0Config(spendFroms, tx0FeeTarget, mixFeeTarget);
-    return whirlpoolWallet.tx0(tx0Config, pool);
+    return whirlpoolWallet.tx0Cascade(tx0Config);
   }
 
   private Collection<WhirlpoolUtxo> findAutoTx0SpendFrom(