+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog for whirlpool-client.
 ## Unreleased
 ### 
 
+- refactor: fix tx0 decoy when spending utxos with same hash or with tx0MaxOutputs limit *(2023-09-26)*
 - refactor: count tx0 decoy change for fee estimation + add Tx0PreviewConfig.decoyTx0x2Forced *(2023-09-26)*
 - refactor: add utxo abstraction with UtxoRef, UtxoDetail, UtxoSpendable *(2023-09-24)*
 - add Tx0Preview.decoyTx0x2 *(2023-09-16)*